# api/openapi/ledgrionyx-v1-openapi.yaml
openapi: 3.0.3
info:
  title: Ledgrionyx API
  version: 1.0.0
  description: >
    Ledgrionyx is an institution-grade financial accounting API designed for
    multi-tenant organizations, ledger-driven financial operations, and
    enterprise integrations.
servers:
  - url: https://api.ledgrionyx.com/v1
    description: Production
  - url: https://sandbox.api.ledgrionyx.com/v1
    description: Sandbox
security:
  - OAuth2: []
...
paths:
  /auth/token:
    post:
      tags: [Authentication]
      summary: Exchange client credentials for an access token
      operationId: createAccessToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, client_secret, grant_type]
              properties:
                client_id:
                  type: string
                client_secret:
                  type: string
                grant_type:
                  type: string
                  enum: [client_credentials]
      responses:
        '200':
          description: Access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: Bearer
                  expires_in:
                    type: integer
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /auth/cli-login:
    post:
      tags: [Authentication]
      summary: Exchange a CLI API key for a short-lived bearer token
      operationId: cliLogin
      security: []
      servers:
        - url: https://api.ledgrionyx.com
          description: Production root
        - url: https://sandbox.api.ledgrionyx.com
          description: Sandbox root
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CLIAuthLoginRequest'
      responses:
        '200':
          description: CLI session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CLIAuthSessionResponse'
        '400':
          $ref: '#/components/responses/StandardBadRequest'
        '401':
          $ref: '#/components/responses/StandardUnauthorized'
        '404':
          $ref: '#/components/responses/StandardNotFound'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /auth/token
          scopes:
            auth:manage: Manage API credentials
            org:read: Read organization metadata
            org:write: Create or modify organization metadata
            team:read: Read roles, permissions, and team members
            team:write: Invite or deactivate team members
            migration:read: Read migration jobs and results
            migration:write: Create or execute migration workflows
            accounts:read: Read chart of accounts data
            accounts:write: Create or modify accounts
            customers:read: Read customer records
            customers:write: Create or modify customers
            vendors:read: Read vendor records
            vendors:write: Create or modify vendors
            invoices:write: Create invoices
            bills:read: Read bills
            bills:write: Create bills
            payments:write: Record invoice or bill payments
            ledger:read: Read journal entries
            ledger:write: Post journal entries or ledger-affecting migrations
            banking:read: Read bank accounts, transactions, and reconciliations
            banking:write: Import bank activity and create reconciliations
            reports:read: Read financial reports
            webhooks:read: Read webhook endpoints and delivery logs
            webhooks:write: Create webhook endpoints or replay events
            audit:read: Read retained system events and audit-facing logs
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    StandardBadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardError'
    StandardUnauthorized:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardError'
...
    StandardNotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardError'
...
  schemas:
...
    CLIUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
...
    CLIAuthLoginRequest:
      type: object
      required: [api_key, organization_id]
      properties:
        api_key:
          type: string
        organization_id:
          type: string
...
    CLIAuthSessionResponse:
      type: object
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
        organization_id:
          type: string
        user:
          $ref: '#/components/schemas/CLIUser'
...
    StandardError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - INSUFFICIENT_SCOPE
                - NOT_FOUND
                - METHOD_NOT_ALLOWED
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: object
              additionalProperties: true
      example:
        error:
          code: INSUFFICIENT_SCOPE
          message: "Missing required API scope(s): reports:read"
          details: {}
...